Removed arc+mutex from OpenStdFile.
This requires transferring ownership of the files, but it also means
that the locking is purely handled by the transaction. Previously the
isolation wasn't actually quite right due to the order of releasing the
locks.

Now if there's ever a bug with improper isolation, it will yield a panic
when trying to write to the file. If this happens, look to how the
transaction handle is being released before the file is being released.