Fixed bad caching with file id 0
When switching from Paths to IDs in the caches, I didn't properly handle
this case. I decided to not cache all chunks, but then I still gave a
potential key ID (the file ID) to the file, which the chunk cache would
try to look up results from.

Well, what happens when you call initialize_state on two separate files
and they happen to contain headers in the page offsets in the file? If
read_chunk previously cached a value for another file (and stored it at
file id 0), the next initialize_state that requested that same data
would get data from another file.

By properly using None to denote no file ID, we avoid the issue.

This didn't affect unit tests in this project, but rather it affected
unit tests in BonsaiDb.