Made LogEntry accept arbitrary bytes
Fixes #14.

This allows BonsaiDb to use its existing structures, and will allow any
other users of Nebari to use it however they please. Nebari only needs
the transaction ID to perform all of its ACID guarantees, the remainder
of the log entry space can be used to store metadata related to the
transaction.