Removed bad logic
This was added when I was trying to debug the original tree balancing
code. The reason it's not necessary is that absorbing only happens after
a node has fully processed all the modifications that apply to it. Thus,
when we merge the node, we no longer need to re-evaluate the previous
node.

The result wasn't a failure, just that it re-evaluated a node it didn't
need to.