Refactored dynamic_order
Now that trees balance properly, it makes more sense to tighten the
order back up. The idea of the cube root is to try to grow the tree to a
depth of 3 with a slowly increasing MAX_ORDER.

This doesn't work perfectly though, because the root doesn't enforce a
certain density before allowing a split. But, it works well enough that
reducing this horrible hack of a heruistic to this much simpler one is a
net performance increase in almost all categories in initial testing.