Refactored ChangeResult to not pass nodes
The receiver of a ChangeResult is now responsible for extracting the
nodes itself.

Additionally, we now have another mechanism for tree balancing: moving
nodes from a node that needs to split into its neighbor to avoid the
split. We currently only move nodes to a previous child, but that's the
one I was mostly aiming to implement -- for sequential IDs, the trees
generated were very unbalanced yielding more hops on disk.

Locally, these changes are a little bit of a mixed-bag performance wise.
They implact the dynamic_order's effectiveness, but we haven't updated
the algorithm. The more tightly packed the nodes, the more time is spent
inside of each node looking for an element. Thus, it impacts how large
we want to allow each node to be under each given dataset size.

So, for now, benchmarks will be a bit up and down -- some drastically
up, most unchanged, some moderately slower. I'm confident at the end of
the day, this is moving in the right direction.