Added embedded indexes
Fixes #21, #19

The scan API has been expanded to include a node callback, and the
indexes are always passed in now.

The indexes now contain an `embedded` field, which contains the
EmbeddedStats generic parameter. When a value is saved, updated, or
deleted, `EmbeddedIndex::index` is executed. As each node is saved in
the tree and the internal stats are reduced, so is the embedded index.

The example is a bit contrived, but shows off the power: you can see how
the interior node logging lines contain the sum of all the nodes
inside of it.