Allowed truncating a file to empty.
This sounds like a scary proposition, which is why I coded it
defensively originally. However, I ran into a real-world use case in
BonsaiDb: if an app aborts the first transaction that's being written to
a file, then closes the file and reopens it, the file will throw an
error rather than assuming a previously empty file.

My original thought was to always write out the header when initializing
the file, but this fix seemed simpler.