Fixed a deadlock with transactions
I was listening to a podcast where the author was talking about PostgreSQL and how it achieved MVCC, and something in the way locks were being discussed made me question if I had actually thought through the problem of dependent locks.

This change introduces an allocation, but it's the best change in my opinion. An alternative would be to document the requirement for tree names to be ordered on `new_transaction` but that feels worse to me. If it actually shows up in profiling as being significant, we can introduce separate methods that allow you to bypass that allocation and sort.