Fixing and testing compaction of MemoryFiles
The MemoryFileManager uses weak references to buffers to allow things to
clean up automatically when the last reference goes away. `replace_with`
was previously incompatible with that idea, as compaction created a
file, wrote to it, closed it, and returned the path to the file. This
left no references to the underlying buffer.

My changing replace_with to take the file, it solves all the problems.
The rest of the fixes are just subtle ID management issues that the
MemoryFileManager was ignoring -- and weren't important until compaction
worked.