Fix modify() edge case with larger trees
I ran into this while working on a view that had 70k entries in it. I
changed the view version, which caused the mapper to re-map the entire
view. BonsaiDb currently overwrites the existing data (although very
shortly that will change to file deletions), which was causing a
significant number of changes to be happening in bulk CompareSwap
operations.

I hit the debug_assert that children nodes got out of order. I ended up
hand-writing a test that tried to simulate the setup that I had in
BonsaiDb, and eventually I came up with a test that was able to
reproduce the debug_assert.

Ideally, this sort of bug could have been found using Fuzzing. The
problem I see with this sort of bug is that it requires the database to
be large enough, and for a batch operation to perform a set of
operations that breaks the database. Still, it's worth trying to find
more bugs using as many approaches as possible. Refs #50.