Fixed absborb edge case and added paranoid feature
This absorb edge case was found with fuzzing. The previous fix bubbled
up the status of should_backup, but that wasn't a complete picture of
the state of the tree. So, the reproduction scenario is almost the same
-- except an even narrower set of circumstances where we needed to back
up but weren't returning the correct status.

This change just always returns true when we absorb. It's the safest
option, and my initial attempt at correct logic still failed the test
case. Thus, I reasoned if I couldn't figure out what the value should
be, I should just always return true. The worse case is that this causes
one extra key comparison -- but that's only when we're in the middle of
absorbing, and after thinking about this all day while the fuzzer ran, I
think it might be the only way for this function to work correctly
without passing it additional inforamtion.

The paranoid feature is for fuzzing. When disabled, Nebari only performs
expensive tree sanity checks in debug builds. In release builds, the
sanity checks are removed. The paranoid feature enables the sanity
checks even in release builds. This is important for the fuzzer, since
we want it to crash when the tree is put in an invalid state -- but for
fuzzing, we want to run in release mode to run through as many
possibilities as possible.