Exposing more B+Tree functionality
This largely enables external Roots to be written. As part of
khonsulabs/bonsaidb#250, I am moving the document map into the view
entries tree, which requires using two roots like the Versioned root.
The biggest limiting factor was a lot of unexported functionality and
types.

This set of changes also removes some of the &mut requirements for some
of the closures.

The last major change is adding the Value associated type to Root, which
allows a tree to use something other than ArcBytes on its public API
surface. The two built-in trees will continue to use ArcBytes, but the
ViewEntries tree in BonsaiDb will be using a custom type to prevent
extra deserialization.